task1/requirements.txt
blis==0.2.4
certifi==2020.11.8
chardet==3.0.4
ChatterBot==1.0.8
chatterbot-corpus==1.2.0
click==7.1.2
cymem==2.0.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.7.3
joblib==0.17.0
MarkupSafe==1.1.1
mathparse==0.1.2
murmurhash==1.0.4
nltk==3.5
numpy==1.19.4
plac==0.9.6
preshed==2.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
PyYAML==3.13
regex==2020.11.11
requests==2.25.0
six==1.15.0
spacy==2.1.9
SQLAlchemy==1.3.20
srsly==1.0.4
thinc==7.0.8
tqdm==4.51.0
urllib3==1.26.1
wasabi==0.8.0
Werkzeug==1.0.1
